DL-2048 No exception when skipping to review details page
